% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{miniMOOCApp}
\alias{miniMOOCApp}
\title{Create a miniMOOC shinyEvents app}
\usage{
miniMOOCApp(mm = readRDS("mm.Rds"), log.file = NULL, title = NULL,
  window.title = title)
}
\arguments{
\item{mm}{A compiled miniMOOC object created with \code{parse_mooc_rmd}. You can save the result of \code{parse_mooc_rmd} with \code{saveRDS} and load it with \code{readRDS}.}

\item{log.file}{an optional log file in which answers to multiple choice quizzes are anomyously logged in csv format.}

\item{title}{Optional title shown above the sections tabSet.}

\item{window.title}{title of browser window}
}
\description{
Create a miniMOOC shinyEvents app
}
