% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{parse_mooc_rmd}
\alias{parse_mooc_rmd}
\title{Parse and compile an Rmd}
\usage{
parse_mooc_rmd(file, chunks = c("knit", "render", "ignore")[1],
  youtube.width = 560, youtube.height = round((315/560) * youtube.width),
  lang = "en", left.margin = 1, right.margin = left.margin)
}
\arguments{
\item{file}{Rmd file path.}

\item{chunks}{what shall be done with R code chunks. By default "knit", the other option "render" allows some alternative behavior not yet well documented.}

\item{youtube.width}{default width of included youtube iframes. By default 560 you can also set 720.}

\item{youtube.height}{default height of included youtube iframes. By default keeps Google's proposed aspect ratio to width.}

\item{lang}{default language for quiz messages. Currently only "en" english and "de" German supported.}
}
\description{
Parse and compile an Rmd
}
