% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{preview_mooc_rmd}
\alias{preview_mooc_rmd}
\title{Preview an Rmd file as miniMOOC page}
\usage{
preview_mooc_rmd(file, log.file = NULL, title = NULL,
  window.title = title, ...)
}
\description{
You can essentially pass all arguments you pass to miniMOOC app and parse_mooc_rmd.
}
