% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rmd.R
\name{youtube.hashdot.to.iframe}
\alias{youtube.hashdot.to.iframe}
\title{Converts the short tag for a youtube video}
\usage{
youtube.hashdot.to.iframe(str, width = 560, height = round(width * 0.5625))
}
\description{
#. youtube id = "youtubeid"
}
\details{
To proper HTML code that embeds the iframe with given width and height in pixels.
}
